WIP: render() and render_buffer
The `render()` hook needs to be called on a view if present. The
passed buffer should also be respected if strings or DOM are pushed
or if a view is appended.

This gets the `render` hook to be called, but requires further work
for the render_buffer to work correctly. My current thinking is that
the render buffer can return a partially executed template (one with
a cachedFragment that is its DOM from the string or pushed nodes) and
a hydration path that attaches any views.

WIP testing at:
http://localhost:4200/tests/?module=Metamorph%20views&forceskip