Introduce test-resolver, migrate tests
Test in Ember have traditionally depended on the Ember.DefaultResolver,
also known as the Ember "globals" resolver. Despite this, Ember
applications have already moved again from the default resolver to the
resolver provided by the ember-resolver NPM package.

The globals resolver is legacy code at this point, mostly used for demo
purposes. As we iterate on the DI implementation in Ember, we want to
move away from coupling to this code.

Thus this PR introduced a test resolver which should replace the default
resolver throughout Ember's test suite.