Introduce source to injections
Ember templates use a `source` argument to the container. This hints
where to look for "local lookup", and also what namespace to look in for
a partial specifier. For example a template in an addon's `src` tree can
invoke components and helpers in that tree becuase it has a `source`
argument passed to lookups.

This introduces the same functionality for service injections. A service
can now accept a `source` argumnet and its lookup will apply to the
namespace of that file. As with templates, this is intended to be a
build-time feature. We can run an AST transform on JS to add this
`source` argumnet to any injections.